remove ResizableWindow class and use toolbar for dragging
